chore: move fn
